+++ service/controller/resource/cloudconfig/error.go
@@ -1,9 +1,23 @@
 package cloudconfig
 
 import (
+	"strings"
+
 	"github.com/Azure/go-autorest/autorest"
 	"github.com/giantswarm/microerror"
 )
+
+// executionFailedError is an error type for situations where Resource
+// execution cannot continue and must always fall back to operatorkit.
+//
+// This error should never be matched against and therefore there is no matcher
+// implement. For further information see:
+//
+//     https://github.com/giantswarm/fmt/blob/master/go/errors.md#matching-errors
+//
+var executionFailedError = &microerror.Error{
+	Kind: "executionFailedError",
+}
 
 var invalidConfigError = &microerror.Error{
 	Kind: "invalidConfigError",
@@ -49,4 +63,24 @@ var wrongTypeError = &microerror.Error{
 // IsWrongTypeError asserts wrongTypeError.
 func IsWrongTypeError(err error) bool {
 	return microerror.Cause(err) == wrongTypeError
+}
+
+// IsStorageAccountNotProvisioned asserts storage account not provisioned error from upstream's API message.
+func IsStorageAccountNotProvisioned(err error) bool {
+	if err == nil {
+		return false
+	}
+	return strings.Contains(microerror.Cause(err).Error(), "StorageAccountIsNotProvisioned")
+}
+
+var tooManyCredentialsError = &microerror.Error{
+	Kind: "tooManyCredentialsError",
+}
+
+var missingOrganizationLabel = &microerror.Error{
+	Kind: "missingOrganizationLabel",
+}
+
+var missingReleaseVersionLabel = &microerror.Error{
+	Kind: "missingReleaseVersionLabel",
 }